* Added ResponseArray codable * Updated repository main layer

+++ Movie/Model/Movie.swift
@@ -13,7 +13,7 @@ class Movie: Codable {
     var id: Int?
     var vote_count: Int?
     var video: Bool?
-    var vote_average: Int?
+    var vote_average: Double?
     var title: String?
     var popularity: Double?
     var poster_path: String?
